requirements.txt
python-docx==0.8.11
pandas==2.0.3
openpyxl==3.1.2 